backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.1.0
django-elasticsearch-dsl==7.3
django-filter==23.2
django-redis==5.2.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Faker==18.11.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.0
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyaes==1.6.1
PyJWT==2.7.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-django==4.5.2
pytest-env==0.8.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
